% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bullpen}
\alias{bullpen}
\title{Bullpen Information}
\format{

}
\usage{
bullpen
}
\description{
Information on venirepersons who were not disqualified or exempted.
}
\details{
Note: this dataset was compiled manually from notes taken during jury
selection. The data is provided here for reference only and may contain
errors.
}
\keyword{datasets}
