% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_odds.R
\name{calculate_odds}
\alias{calculate_odds}
\title{Calculates odds ratios for a logistic regression model}
\usage{
calculate_odds(model)
}
\arguments{
\item{model}{A logistic regression model}
}
\value{
A table of odds ratios for the coefficients of the logistic regression model
}
\description{
Calculates odds ratios for a logistic regression model
}
\examples{
calculate_odds(lr_caused)
calculate_odds(lr_race)
}
