% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-risk.R
\name{calculate_risk}
\alias{calculate_risk}
\title{Create a relative risk ratio table}
\usage{
calculate_risk(x)
}
\arguments{
\item{x}{A logistic regression model.}
}
\value{
A risk ratio table.
}
\description{
Create a relative risk ratio table
}
\examples{
calculate_risk(lr_caused)
}
