% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coefficient_positions.R
\name{get_coef_positions}
\alias{get_coef_positions}
\title{Get coefficient positions to assist with calling by number}
\usage{
get_coef_positions(model)
}
\arguments{
\item{model}{A logistic regression model}
}
\value{
A table of coefficients with their respective line numbers for a
    logistic regression model
}
\description{
Get coefficient positions to assist with calling by number
}
\examples{
get_coef_positions(lr_caused)
get_coef_positions(lr_race)
}
