% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juror-package.R
\docType{package}
\name{juroR-package}
\alias{juroR}
\alias{juroR-package}
\title{juroR: Jury Data for Felony Trials in Dallas County}
\description{
This package contains data used in the paper: "Examining Bias in Jury Selection for Criminal Trials in Dallas County"
}
\keyword{internal}
