% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juror-package.R
\docType{package}
\name{juryR-package}
\alias{juryR}
\alias{juryR-package}
\title{juryR: Jury Data for Felony Trials in Dallas County}
\description{
This package contains data used in the paper: "Examining Bias in Jury Selection for Criminal Trials in Dallas County"
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mattfarrow1.github.io/juryr/}
}

}
\keyword{internal}
