% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{lr_bias_a_g_def}
\alias{lr_bias_a_g_def}
\title{Bias against the defense, by race.}
\format{
An object of class \code{glm} (inherits from \code{lm}) of length 30.
}
\usage{
lr_bias_a_g_def
}
\description{
A logistic regression model whose response variable is `bias_a_g_def` and
whose explanatory variable is `jur_race_w`.
}
\keyword{datasets}
