% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{lr_bias_a_g_def_g}
\alias{lr_bias_a_g_def_g}
\title{Bias against the defense, by gender.}
\format{
An object of class \code{glm} (inherits from \code{lm}) of length 30.
}
\usage{
lr_bias_a_g_def_g
}
\description{
A logistic regression model whose response variable is `bias_a_g_def` and
whose explanatory variable is `jur_gend`.
}
\keyword{datasets}
