% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{lr_cause_5am_a}
\alias{lr_cause_5am_a}
\title{Caused by inability to accept a defendant's 5th Amendment rights, by age
group and I-30 split.}
\format{
An object of class \code{glm} (inherits from \code{lm}) of length 31.
}
\usage{
lr_cause_5am_a
}
\description{
A logistic regression model whose response variable is `ammend_5` and whose
explanatory variables are `i_30` and `age_group`.
}
\keyword{datasets}
