% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{lr_cause_5am_g}
\alias{lr_cause_5am_g}
\title{Caused by inability to accept a defendant's 5th Amendment rights, by gender.}
\format{
An object of class \code{glm} (inherits from \code{lm}) of length 30.
}
\usage{
lr_cause_5am_g
}
\description{
A logistic regression model whose response variable is `ammend_5` and whose
explanatory variable is `jur_gend`.
}
\keyword{datasets}
