% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{lr_cause_bop}
\alias{lr_cause_bop}
\title{Burden of proof, by race.}
\format{
An object of class \code{glm} (inherits from \code{lm}) of length 30.
}
\usage{
lr_cause_bop
}
\description{
A logistic regression model whose response variable is `bop_100_percent` and
whose explanatory variable is `jur_race_w`.
}
\keyword{datasets}
