% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{lr_cause_bop_a}
\alias{lr_cause_bop_a}
\title{Burden of proof, by age group.}
\format{
An object of class \code{glm} (inherits from \code{lm}) of length 30.
}
\usage{
lr_cause_bop_a
}
\description{
A logistic regression model whose response variable is `bop_100_percent` and
whose explanatory variable is `age_group`.
}
\keyword{datasets}
