% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{lr_cause_inn}
\alias{lr_cause_inn}
\title{Presumption of innocence, by race and I-30 split.}
\format{
An object of class \code{glm} (inherits from \code{lm}) of length 31.
}
\usage{
lr_cause_inn
}
\description{
A logistic regression model whose response variable is `inn` and
whose explanatory variables are `i_30` and `jur_race_w`.
}
\keyword{datasets}
