% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{lr_cause_pun_a}
\alias{lr_cause_pun_a}
\title{Inability to accept the full range of punishment, by age group.}
\format{
An object of class \code{glm} (inherits from \code{lm}) of length 30.
}
\usage{
lr_cause_pun_a
}
\description{
A logistic regression model whose response variable is `jur_pun` and whose
explanatory variable is `age_group`.
}
\keyword{datasets}
