% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{lr_caused}
\alias{lr_caused}
\title{Caused, by race.}
\format{
An object of class \code{glm} (inherits from \code{lm}) of length 30.
}
\usage{
lr_caused
}
\description{
A logistic regression model whose response variable is `jur_caused` and whose
explanatory variable is `jur_race_w`.
}
\keyword{datasets}
