% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{lr_caused_a}
\alias{lr_caused_a}
\title{Caused, by age group only.}
\format{
An object of class \code{glm} (inherits from \code{lm}) of length 30.
}
\usage{
lr_caused_a
}
\description{
A logistic regression model whose response variable is `jur_caused` and whose
explanatory variable is `age_group`.
}
\keyword{datasets}
