% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{lr_caused_i30}
\alias{lr_caused_i30}
\title{Caused by I-30 split and race.
A logistic regression model whose response variable is `jur_caused` and whose
explanatory variables are `i_30` and `jur_race_w`.}
\format{
An object of class \code{glm} (inherits from \code{lm}) of length 31.
}
\usage{
lr_caused_i30
}
\description{
Caused by I-30 split and race.
A logistic regression model whose response variable is `jur_caused` and whose
explanatory variables are `i_30` and `jur_race_w`.
}
\keyword{datasets}
