% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{lr_i30}
\alias{lr_i30}
\title{Caused by I-30 split only.}
\format{
An object of class \code{glm} (inherits from \code{lm}) of length 31.
}
\usage{
lr_i30
}
\description{
A logistic regression model whose response variable is `jur_caused` and whose
explanatory variable is `i_30`.
}
\keyword{datasets}
