% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{lr_pun_min_a}
\alias{lr_pun_min_a}
\title{Inability to accept the minimum punishment, by age group.}
\format{
An object of class \code{glm} (inherits from \code{lm}) of length 30.
}
\usage{
lr_pun_min_a
}
\description{
A logistic regression model whose response variable is `pun_min` and whose
explanatory variable is `age_group`.
}
\keyword{datasets}
