% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{venire}
\alias{venire}
\title{Venire Information}
\format{

}
\usage{
data(venire)
}
\description{
Information on venirepersons who took part in jury selection for the trials
in the study.
}
\keyword{datasets}
