% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bullpen}
\alias{bullpen}
\title{Bullpen Information}
\format{

}
\usage{
data(bullpen)
}
\description{
Information on venirepersons who were not disqualified or exempted.
}
\examples{
data(bullpen)
}
\keyword{datasets}
