% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-odds.R
\name{calculate_odds}
\alias{calculate_odds}
\title{Create an odds ratio table}
\usage{
calculate_odds(x)
}
\arguments{
\item{x}{A logistic regression model.}
}
\value{
An odds ratio table.
}
\description{
Create an odds ratio table
}
\examples{
calculate_odds(lr_caused)
}
