% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient-positions.R
\name{get_coef_positions}
\alias{get_coef_positions}
\title{Get the position numbers of the coefficients}
\usage{
get_coef_positions(x)
}
\arguments{
\item{x}{A logistic regression model.}
}
\value{
Positions of each coefficient.
}
\description{
Get the position numbers of the coefficients
}
\examples{
get_coef_positions(lr_caused)
}
